Pass the unescaped project id to all usages of ProjectExtensions.FindProject()
FindProject() compares the specified ProjectId value to the namespace and project names of mock projects.
This was failing because most usages were passing the value of ProjectId.ValueAsUriParameter() which is the URL-escaped form of the project path.
When the project path contained a "/", no matching project could be found, causing a NullReferenceException when trying to get a project client from the GitLabClient mock.