Pass the unescaped group id to all usages of GroupExtensions.FindGroup()
FindGroup() compares the specified GroupId value to the namespace and names of mock groups.
This was failing because most usages were passing the value of GroupId.ValueAsUriParameter() which is the URL-escaped form of the project path.
When the group path contained a "/", no matching group could be found, causing a NullReferenceException when trying to get a group client from the GitLabClient mock.