Further tweak dependabot.yml to resolve warnings (#790)
Address the following warnings:

Dependabot couldn't find a Dockerfile
Dependabot requires a Dockerfile to evaluate your Docker dependencies. It had expected to find one at the path: /Dockerfile.

If this isn't a Docker project, you may wish to disable updates for it in the .github/dependabot.yml config file in this repo.

Dependabot couldn't find a package.json
Dependabot requires a package.json to evaluate your JavaScript dependencies. It had expected to find one at the path: /package.json.

If this isn't a JavaScript project, you may wish to disable updates for it in the .github/dependabot.yml config file in this repo.

Co-authored-by: Thomas Cortes <78750681+Toa741@users.noreply.github.com>