Add support for the "Get raw file" API (#897)
* Add support for the "Get raw file" API

Add support for downloading the content's of a file via GitLabs "Get raw file" API.
In contrast to the existing "Get" method that gets a file's content and its metadata, the "GetRaw" method gives access to the file stream.
This avoids having to loads the contents of the entire file into memory.
Additionally, this API allows getting the content of the file if the file was commited via Git LFS ("Get()" will just return the pointer to the LFS file in this case)

GitLab API docs for "Get raw file from repository": https://docs.gitlab.com/api/repository_files/#get-raw-file-from-repository

* Apply suggestions from code review

Co-authored-by: Louis Zanella <louis.zanella@gmail.com>

* Make getting raw file content async

Replace IFilesClient.GetRaw() method with GetRawAsync()

---------

Co-authored-by: Louis Zanella <louis.zanella@gmail.com>